MIsc PR adjustments; cleanup + reformat.
