+++ src/config/docs/ReportsSchema.yml
@@ -1,0 +1,33 @@
+components:
+  schemas:
+    GetSanctionedSellersReportRs:
+      type: array
+      items:
+        type: object
+        properties:
+          seller_id:
+            type: string
+            example: 0d367ba3-a2e8-4380-86c3-ab7c0b7890c0
+          name:
+            type: string
+            example: Test Seller
+          address:
+            type: string
+            example: 1234 Test St, Test City, Democratic People's Republic of Korea
+          sanctioned_location:
+            type: string
+            example: North Korea
+          sell_map_center:
+            type: object
+            properties:
+              type:
+                type: string
+                example: Point
+              coordinates:
+                type: array
+                items:
+                  type: number
+                example: [-73.856077, 40.848447]
+            required:
+              - type
+              - coordinates
