+++ src/models/enums/restrictedAreas.ts
@@ -1,0 +1,9 @@
+export enum RestrictedAreas {
+  CUBA = "Cuba",
+  NORTH_KOREA = "North Korea",
+  IRAN = "Iran",
+  SYRIA = "Syria",
+  REPUBLIC_OF_CRIMEA = "Republic of Crimea",
+  DONETSK_OBLAST = "Donetsk Oblast",
+  LUHANSK_OBLAST = "Luhansk Oblast"
+}
