+++ src/services/report.service.ts
@@ -1,12 +1,11 @@
-import {env} from "../utils/env";
-import logger from "../config/loggingConfig";
-import {ISeller} from "../types";
-import Bottleneck from "bottleneck";
-import {getAllSellers} from "./seller.service";
 import * as Sentry from "@sentry/node";
-import {isInSanctionedRegion, reverseLocationDetails} from "../helpers/sellerReportHelpers";
+import Bottleneck from "bottleneck";
 
-Sentry.init({dsn: env.SENTRY_DSN});
+import {getAllSellers} from "./seller.service";
+import {reverseLocationDetails, isRestrictedLocation} from "../helpers/location";
+import {ISeller} from "../types";
+import logger from "../config/loggingConfig";
+
 const requestLimiter = new Bottleneck({minTime: 1000});
 const processedSellersCache = new Set();
 let sanctionedSellers: ISeller[] = [];
@@ -24,10 +23,8 @@ export const filterSanctionedSellers = async (seller: ISeller) => {
     if (response.data.error) {
       const errorMessage = response.data.error;
       logger.error(`Geocoding error for seller ${seller_id} with original coordinates: ${errorMessage}`);
-
       logger.info(`Retrying geocoding with swapped coordinates for seller ${seller_id}.`);
       response = await reverseLocationDetails(longitude, latitude);
-
       if (response.data.error) {
         const retryErrorMessage = response.data.error;
         logger.error(`Geocoding retry error for seller ${seller_id} with swapped coordinates: ${retryErrorMessage}`);
@@ -39,14 +36,10 @@ export const filterSanctionedSellers = async (seller: ISeller) => {
 
     const locationData = response.data;
     const locationName = locationData.display_name;
-    logger.info(
-      `Seller location: ${locationName}, from coordinates: latitude: ${latitude}, longitude: ${longitude}`
-    );
+    logger.info(`Seller location: ${locationName}, from coordinates: latitude: ${latitude}, longitude: ${longitude}`);
 
-    if (isInSanctionedRegion(locationName)) {
-      logger.warn(
-        `Pioneer is selling in a restricted area | ${seller_id} | ${name} | ${latitude} | ${longitude} | ${locationName}`
-      );
+    if (isRestrictedLocation(locationName)) {
+      logger.warn(`Pioneer is selling in a restricted area | ${seller_id} | ${name} | ${latitude} | ${longitude} | ${locationName}`);
       sanctionedSellers.push(seller);
     }
 
@@ -56,15 +49,13 @@ export const filterSanctionedSellers = async (seller: ISeller) => {
   }
 };
 
-
-
 export const reportSanctionedSellers = async () => {
   const sellers = await getAllSellers();
   sanctionedSellers = [];
   logger.info(`Total number of sellers: ${sellers.length}`);
   await processSellers(sellers);
-  if(sanctionedSellers.length > 0){
-    logger.info(`Total number of blacklisted sellers: ${sanctionedSellers.length}`)
+  if (sanctionedSellers.length > 0) {
+    logger.info(`Total number of blacklisted sellers: ${sanctionedSellers.length}`);
     logger.info("Generating Weekly Report on Sanctioned Sellers");
     Sentry.captureMessage("Weekly Sanctioned Sellers Report", {
       level: "info",
@@ -73,7 +64,7 @@ export const reportSanctionedSellers = async () => {
     logger.error(`Weekly report data on Sanctioned Sellers: ${sanctionedSellers}`);
   }
 
-  if(geocodingErrors.length > 0){
+  if (geocodingErrors.length > 0) {
     logger.info(`Total number of sellers with geocoded error locations: ${geocodingErrors.length}`);
     Sentry.captureMessage("Failed to geocode Sellers location", {
       level: "info",
@@ -93,4 +84,3 @@ const processSellers = async (sellers : ISeller[]) => {
 }
 
 const fetchSellerLocation = requestLimiter.wrap(filterSanctionedSellers);
-
